streaming/pyproject.toml
[project]
name = "streaming"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.6.0",
    "google-generativeai>=0.8.5",
    "openai-agents>=0.1.0",
    "python-dotenv>=1.1.1",
    "sdk>=1.0.0",
]
